pyproject.toml
[project]
name = "pycaret-ng"
version = "1.0.0"
description = "pycaret-ng — modernized soft-fork of PyCaret (low-code ML for Python 3.9–3.13, sklearn >=1.6, pandas >=2.2, modern sktime/matplotlib/yellowbrick)."
readme = "README.md"
keywords = [
  "data-mining",
  "data-science",
  "classification",
  "clustering",
  "detection",
  "forecasting",
  "machine-learning",
  "regression",
  "scikit-learn>=1.6,<2",
  "time-series",
]
license = { file = "LICENSE" }
maintainers = [
  { name = "pycaret developers" },
]
# for full credits see contributors.md
authors = [
    { name = "pycaret developers" },
]
requires-python = ">=3.10"
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX",
  "Operating System :: Unix",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development",
]
# core dependencies of pycaret
# this set should be kept minimal!
dependencies = [
  # Base
  "ipython>=5.5.0",
  "ipywidgets>=7.6.5",
  "tqdm>=4.62.0",
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "jinja2>=3",
  "scipy>=1.11,<2",  # 1.11.4 had no Python 3.13 wheel; lifted in Phase 5 release hygiene.
  "joblib>=1.4.2,<1.5",
  "scikit-learn>=1.6,<2",
  "pyod>=1.1.3",
  "imbalanced-learn>=0.14.0,<0.15.0",
  "category-encoders>=2.7,<3",
  "lightgbm>=3.0.0",
  "numba>=0.55.0",
  "requests>=2.27.1",  # Required by pycaret.datasets
  "psutil>=5.9.0",
  "markupsafe>=2.0.1",  # Fixes Google Colab issue
  "importlib_metadata>=4.12.0",
  "nbformat>=4.2.0",
  "cloudpickle",
  "deprecation>=2.1.0",
  "xxhash",
  "wurlitzer; platform_system != 'Windows'",
  "trio>=0.22.0",  # cap removed: original workaround was for httpcore <1.0; modern httpcore handles trio cleanly. Lift unblocks trio on Python 3.13.
  "setuptools; python_version>='3.12'",
  # Plotting
  "matplotlib>=3.8",
  "mljar-scikit-plot",
  "yellowbrick>=1.4",
  "plotly>=5.14.0",
  "kaleido>=0.2.1",
  "schemdraw>=0.16",
  "plotly-resampler>=0.8.3.1",
  # Time-series
  "statsmodels>=0.14,<1",
  "sktime>=0.31",
  "tbats>=1.1.3",
  "pmdarima>=2.0.4",
]

[project.optional-dependencies]
# there are the following dependency sets:
# - full - all soft dependencies
# - analysis
# - models
# - tuners
# - mlops
# - parallel
# - dev
# - test

# soft dependencies are not required for the core functionality of pycaret

# all soft dependencies
#
# users can install via "pip install pycaret[full]"
#
full = [
  "shap>=0.47",  # 0.46 colorconv:819 calls np.dtype(np.floating) which numpy 2 rejects.
  "interpret>=0.2.7",
  "umap-learn>=0.5.2",
  "pyyaml",
  "ydata-profiling>=4.18",  # Pre-4.18 requires htmlmin which fails to build on Python 3.13 (cgi module removed).
  "explainerdashboard>=0.3.8",  # For dashboard method
  "fairlearn==0.7.0",  # For check_fairness method
  "xgboost>=2.0.0",
  "catboost>=1.2.5",
  "kmodes>=0.11.1",
  "mlxtend>=0.19.0",
  "statsforecast<1.8.0,>1.6",
  "scikit-learn-intelex>=2025.0; platform_machine == 'x86_64' or platform_machine == 'AMD64'",  # 2024.x bundled an older daal4py that imports sklearn.utils._joblib (removed in sklearn 1.6).
  "tune-sklearn",
  "ray[tune]",
  "hyperopt>=0.2.7",
  "optuna>=3.0.0",
  "optuna-integration",
  "scikit-optimize>=0.9.0",
  "mlflow>=2.0.0",
  "gradio>=3.50.2",
  "boto3>=1.24.56",  # For deploy_model method
  "fastapi",  # For web api
  "uvicorn>=0.17.6",  # For web api
  "m2cgen>=0.9.0",  # For model conversion
  "evidently<0.4.30",  # test_check_drift fails with 0.4.30
  "dask<2024.6.3",
  "distributed<2024.6.3",
  "fugue==0.9.1",
  "flask",
  "Werkzeug>=2.2,<3.0",
  "pytest==8.3.2",
  "moto<5.0.0",  # missing moto.mock_s3 on moto>=5.0.0
  # "codecov>=2.1.12",  # https://github.com/home-assistant/core/issues/91283
  "fugue[dask]",
  "dash[testing]",
  "black>=24.8.0",
  "isort>=5.13.2",
  "flake8>=7.1.1",
  "setuptools>=71.1.0",
  "mypy>=0.11.1",
]

analysis = [
  "shap>=0.47",  # 0.46 colorconv:819 calls np.dtype(np.floating) which numpy 2 rejects.
  "interpret>=0.2.7",
  "umap-learn>=0.5.2",
  "pyyaml",
  "ydata-profiling>=4.18",  # Pre-4.18 requires htmlmin which fails to build on Python 3.13 (cgi module removed).
  "explainerdashboard>=0.3.8",  # For dashboard method
  "fairlearn==0.7.0",  # For check_fairness method
]

models = [
  "xgboost>=2.0.0",
  "catboost>=1.2.5",
  "kmodes>=0.11.1",
  "mlxtend>=0.19.0",
  "statsforecast<1.8.0,>1.6",
  "scikit-learn-intelex>=2025.0; platform_machine == 'x86_64' or platform_machine == 'AMD64'",  # 2024.x bundled an older daal4py that imports sklearn.utils._joblib (removed in sklearn 1.6).
]

tuners = [
  "tune-sklearn",
  "ray[tune]",
  "hyperopt>=0.2.7",
  "optuna>=3.0.0",
  "optuna-integration",
  "scikit-optimize>=0.9.0",
]

mlops = [
  "mlflow>=2.0.0",
  "gradio>=3.50.2",
  "boto3>=1.24.56",  # For deploy_model method
  "fastapi",  # For web api
  "uvicorn>=0.17.6",  # For web api
  "m2cgen>=0.9.0",  # For model conversion
  "evidently<0.4.30",  # test_check_drift fails with 0.4.30
]

parallel = [
  "dask<2024.6.3",
  "distributed<2024.6.3",
  "fugue==0.9.1",
  "flask",
  "Werkzeug>=2.2,<3.0",
]

prophet = [
  "prophet>=1.0.1",
]

# dev - the developer dependency set, for contributors to sktime
dev = [
  "black>=24.8.0",
  "isort>=5.13.2",
  "flake8>=7.1.1",
  "setuptools>=71.1.0",
  "mypy>=0.11.1",
]


# CI related soft dependency sets - not for users of pycaret
# they are stable and subject to deprecation policies
test = [
  "pytest==8.3.2",
  "moto<5.0.0",  # missing moto.mock_s3 on moto>=5.0.0
  # "codecov>=2.1.12",  # https://github.com/home-assistant/core/issues/91283
  "fugue[dask]",
  "dash[testing]",
]


[project.urls]
"API Reference" = "https://pycaret.gitbook.io/"
Documentation = "https://pycaret.gitbook.io/"
Download = "https://pypi.org/project/pycaret/#files"
Homepage = "https://pycaret.org/"
"Release Notes" = "https://github.com/pycaret/pycaret/releases"
Repository = "https://github.com/pycaret/pycaret"

[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>61",
]

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]

# uv-specific: declare missing build-time dependencies for transitive
# packages that fail under uv's strict PEP 517 build isolation. pip's
# legacy resolver tolerated these; uv requires them to be explicit.
[tool.uv.extra-build-dependencies]
pmdarima = ["numpy", "setuptools", "cython"]
